requirements.txt
python-docx>=0.8.11
Pillow>=10.0.0
streamlit>=1.28.0
pytest>=7.0.0
